[bug-set-default-theme-active] moved function to themeUtils for setting active theme class, included call in theme selector component to set initial active item
